Add PRIVATE config option
If the site is configured to be private, all pages with content require login.
Non-logged-in users only get to see the login page and static assets.  This is
to make the site suitable for hosting on a public network without leaking
internal data.

Add a login_required_if_private decorator to indicate views that should require
login if PRIVATE is set to True